refactor: remove unused classes
