+++ hedera-mirror-web3/src/main/java/com/hedera/services/jproto/JContractIDKey.java
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.hedera.services.jproto;
 
 import com.hederahashgraph.api.proto.java.ContractID;
@@ -22,15 +21,14 @@ import com.hederahashgraph.api.proto.java.ContractID;
  * Maps to proto Key of type contractID.
  *
  * <p>Just as a public-private key pair is used to control permissions for a Hedera entity (e.g., an
- * account), a {@code contractID} key is <i>also</i> used to control permissions for an entity. The
- * difference is that a {@code contractID} key requires a particular contract to be executing in
- * order to count as having "signed".
+ * account), a {@code contractID} key is <i>also</i> used to control permissions for an entity. The difference is that a
+ * {@code contractID} key requires a particular contract to be executing in order to count as having "signed".
  *
  * <p>For example, suppose {@code 0.0.X} is an account. Its key is {@code Key{contractID=0.0.C}}.
  *
  * <p>Then <b>if</b> we are executing a contract operation, and the EVM receiver address in the
- * current frame is {@code 0.0.C}, then this EVM transaction can use system contracts to, for
- * example, transfer all the hbar from account {@code 0.0.X}.
+ * current frame is {@code 0.0.C}, then this EVM transaction can use system contracts to, for example, transfer all the
+ * hbar from account {@code 0.0.X}.
  */
 public class JContractIDKey extends JKey {
     private final long shardNum;
