+++ hedera-mirror-web3/src/main/java/com/hedera/services/jproto/JDelegatableContractIDKey.java
@@ -13,12 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.hedera.services.jproto;
 
 import com.hederahashgraph.api.proto.java.ContractID;
 
-/** Maps to proto Key of type contractID. */
+/**
+ * Maps to proto Key of type contractID.
+ */
 public class JDelegatableContractIDKey extends JContractIDKey {
     public JDelegatableContractIDKey(final ContractID contractID) {
         super(contractID);
