+++ hedera-mirror-web3/src/main/java/com/hedera/services/jproto/JECDSA_384Key.java
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.hedera.services.jproto;
 
 import com.swirlds.common.utility.CommonUtils;
