+++ hedera-mirror-web3/src/main/java/com/hedera/services/jproto/JEd25519Key.java
@@ -13,14 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.hedera.services.jproto;
 
 import com.hederahashgraph.api.proto.java.Key;
 import com.swirlds.common.utility.CommonUtils;
 import java.util.Arrays;
 
-/** Maps to proto Key of type ed25519. */
+/**
+ * Maps to proto Key of type ed25519.
+ */
 public class JEd25519Key extends JKey {
     public static final int ED25519_BYTE_LENGTH = 32;
