+++ hedera-mirror-web3/src/main/java/com/hedera/services/jproto/JKeyList.java
@@ -13,13 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.hedera.services.jproto;
 
 import java.util.LinkedList;
 import java.util.List;
 
-/** Maps to proto Key of type KeyList. */
+/**
+ * Maps to proto Key of type KeyList.
+ */
 public class JKeyList extends JKey {
     private List<JKey> keys;
