+++ hedera-mirror-web3/src/test/java/com/hedera/services/jproto/JContractAliasKeyTest.java
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.hedera.services.jproto;
 
 import static com.swirlds.common.utility.CommonUtils.unhex;
