+++ hedera-mirror-web3/src/test/java/com/hedera/services/jproto/JContractIDKeyTest.java
@@ -13,14 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.hedera.services.jproto;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 import com.hederahashgraph.api.proto.java.ContractID;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class JContractIDKeyTest {
     @Test
