+++ hedera-mirror-web3/src/test/java/com/hedera/services/jproto/JRSA_3072KeyTest.java
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.hedera.services.jproto;
 
 import static org.junit.jupiter.api.Assertions.*;
