--- hedera-mirror-web3/src/test/java/com/hedera/services/jproto/utils/KeyTree.java
@@ -1,49 +1,0 @@
-/*
- * Copyright (C) 2020-2023 Hedera Hashgraph, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hedera.services.jproto.utils;
-
-import com.hederahashgraph.api.proto.java.Key;
-import org.apache.commons.codec.DecoderException;
-
-import com.hedera.services.jproto.JKey;
-
-public class KeyTree {
-    private final KeyTreeNode root;
-
-    private KeyTree(final KeyTreeNode root) {
-        this.root = root;
-    }
-
-    public static KeyTree withRoot(final NodeFactory rootFactory) {
-        return new KeyTree(KeyTreeNode.from(rootFactory));
-    }
-
-    public Key asKey() {
-        return asKey(KeyFactory.getDefaultInstance());
-    }
-
-    public Key asKey(final KeyFactory factoryToUse) {
-        return root.asKey(factoryToUse);
-    }
-
-    public JKey asJKeyUnchecked() {
-        try {
-            return JKey.mapKey(asKey());
-        } catch (final DecoderException impermissible) {
-            throw new IllegalArgumentException("Key " + asKey() + " should have been decode-able!", impermissible);
-        }
-    }
-}
