+++ hedera-mirror-web3/src/test/resources/bootstrap.yml
@@ -8,3 +8,6 @@ spring:
     baselineVersion: 0
     locations: "filesystem:../hedera-mirror-importer/src/main/resources/db/migration/v1"
     target: latest
+  cloud:
+    kubernetes:
+      enabled: false
